refact(decman-lib): move the encode toolkit
The Daml Value constructors, the extraArgs builders, and
TransferValidity move from decman's action_serializer to
decman_lib::framework::encode, as pub. Integrators cannot implement
DamlProtoEncode without them. decman now depends on decman-lib.
