+++ crates/decman-lib/src/framework/mod.rs
@@ -1,4 +1,6 @@
 //! Integrator-facing, template-agnostic pieces.
 
+pub mod encode;
+
 mod template_id;
 pub use template_id::TemplateId;
